requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.0
Pillow>=10.0
python-decouple>=3.8
psycopg2-binary>=2.9
django-cors-headers>=4.3
python-slugify>=8.0
